Completed documenting and moving into correct packages, classes needed to specify the E-Invoice header
