+++ src/main/java/com/github/siboxd/fatturapa/model/invoiceheader/buyer/CessionarioCommittenteType.java
@@ -1,13 +1,14 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoiceheader.buyer;
 
-import com.github.siboxd.fatturapa.model.invoiceheader.supplier.IndirizzoType;
+import com.github.siboxd.fatturapa.model.invoiceheader.IndirizzoType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
 
 
 /**
- * CessionarioCommittenteType<br>
+ * This block contains the data relating to the transferee or customer of the good/service being billed.
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
@@ -15,12 +16,15 @@ import org.simpleframework.xml.Root;
 @Root(name = "CessionarioCommittenteType")
 public class CessionarioCommittenteType {
 
-    @Element(name = "DatiAnagrafici", required = true)
+    @Element(name = "DatiAnagrafici")
     private DatiAnagraficiCessionarioType datiAnagrafici;
-    @Element(name = "Sede", required = true)
+
+    @Element(name = "Sede")
     private IndirizzoType sede;
+
     @Element(name = "StabileOrganizzazione", required = false)
     private IndirizzoType stabileOrganizzazione;
+
     @Element(name = "RappresentanteFiscale", required = false)
     private RappresentanteFiscaleCessionarioType rappresentanteFiscale;
 
@@ -31,7 +35,7 @@ public class CessionarioCommittenteType {
         return datiAnagrafici;
     }
 
-    public void setDatiAnagrafici(DatiAnagraficiCessionarioType datiAnagrafici) {
+    public void setDatiAnagrafici(final DatiAnagraficiCessionarioType datiAnagrafici) {
         this.datiAnagrafici = datiAnagrafici;
     }
 
@@ -39,7 +43,7 @@ public class CessionarioCommittenteType {
         return sede;
     }
 
-    public void setSede(IndirizzoType sede) {
+    public void setSede(final IndirizzoType sede) {
         this.sede = sede;
     }
 
@@ -47,7 +51,12 @@ public class CessionarioCommittenteType {
         return stabileOrganizzazione;
     }
 
-    public void setStabileOrganizzazione(IndirizzoType stabileOrganizzazione) {
+    /**
+     * <b>Note:</b> To be valued only if the buyer is a person who does not reside in Italy
+     * but who, in Italy, has a permanent establishment through which he carries out his business
+     * (he makes sales of goods or services that are billed)
+     */
+    public void setStabileOrganizzazione(final IndirizzoType stabileOrganizzazione) {
         this.stabileOrganizzazione = stabileOrganizzazione;
     }
 
@@ -55,7 +64,12 @@ public class CessionarioCommittenteType {
         return rappresentanteFiscale;
     }
 
-    public void setRappresentanteFiscale(RappresentanteFiscaleCessionarioType rappresentanteFiscale) {
+    /**
+     * <b>Note:</b> To be valued only if the seller is configured as a non-resident entity that
+     * carries out operations relevant to VAT in Italy, and which makes use of a tax representative
+     * in Italy.
+     */
+    public void setRappresentanteFiscale(final RappresentanteFiscaleCessionarioType rappresentanteFiscale) {
         this.rappresentanteFiscale = rappresentanteFiscale;
     }
