+++ src/main/java/com/github/siboxd/fatturapa/model/invoiceheader/buyer/DatiAnagraficiCessionarioType.java
@@ -1,26 +1,29 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoiceheader.buyer;
 
-import com.github.siboxd.fatturapa.model.invoiceheader.IdFiscaleType;
 import com.github.siboxd.fatturapa.model.invoiceheader.AnagraficaType;
+import com.github.siboxd.fatturapa.model.invoiceheader.IdFiscaleType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
 
 
 /**
- * DatiAnagraficiCessionarioType<br>
+ * It contains the buyer personal data
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
  */
 @Root(name = "DatiAnagraficiCessionarioType")
 public class DatiAnagraficiCessionarioType {
-
+    // TODO: 20/01/2019 make base class for all DatiAnagraficiXXX
     @Element(name = "IdFiscaleIVA", required = false)
     private IdFiscaleType idFiscaleIVA;
+
     @Element(name = "CodiceFiscale", required = false)
     private String codiceFiscale;
-    @Element(name = "Anagrafica", required = true)
+
+    @Element(name = "Anagrafica")
     private AnagraficaType anagrafica;
 
     public DatiAnagraficiCessionarioType() {
@@ -30,7 +33,7 @@ public class DatiAnagraficiCessionarioType {
         return idFiscaleIVA;
     }
 
-    public void setIdFiscaleIVA(IdFiscaleType idFiscaleIVA) {
+    public void setIdFiscaleIVA(final IdFiscaleType idFiscaleIVA) {
         this.idFiscaleIVA = idFiscaleIVA;
     }
 
@@ -38,7 +41,14 @@ public class DatiAnagraficiCessionarioType {
         return codiceFiscale;
     }
 
-    public void setCodiceFiscale(String codiceFiscale) {
+    /**
+     * It serves to provide an additional identification element
+     *
+     * @param codiceFiscale The field, if valued, must contain the tax code of the seller that
+     *                      will be composed of <em>11 numeric characters</em>, if it is a legal person,
+     *                      or <em>16 alphanumeric characters</em>, if it is a natural person.
+     */
+    public void setCodiceFiscale(final String codiceFiscale) {
         this.codiceFiscale = codiceFiscale;
     }
 
@@ -46,7 +56,7 @@ public class DatiAnagraficiCessionarioType {
         return anagrafica;
     }
 
-    public void setAnagrafica(AnagraficaType anagrafica) {
+    public void setAnagrafica(final AnagraficaType anagrafica) {
         this.anagrafica = anagrafica;
     }
